+++ src/apple.c
@@ -22,14 +22,14 @@ static bool IsInsideSnake(Apple apple, Segment *tail) {
 Apple GetNewApple(Snake *snake) {
   Apple newApple = Vector2Zero();
   do {
-    newApple.x = GetRandomValue(0, WIDTH_TILES - 1);
-    newApple.y = GetRandomValue(0, HEIGHT_TILES - 1);
+    newApple.x = (float) GetRandomValue(0, WIDTH_TILES - 1);
+    newApple.y = (float) GetRandomValue(0, HEIGHT_TILES - 1);
   } while (IsInsideSnake(newApple, snake->tail));
   return newApple;
 }
 
 void RenderApple(Apple apple) {
   Vector2 fruitCoords = TiletoCartesian(apple);
-  DrawRectangle(fruitCoords.x, fruitCoords.y, TILE_SIZE - TILE_PADDING,
+  DrawRectangle((int) fruitCoords.x, (int) fruitCoords.y, TILE_SIZE - TILE_PADDING,
                 TILE_SIZE - TILE_PADDING, RED);
 }
